refactor: moving types used at higher level up to lib folder
